backend/requirements-cloud.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Google Cloud Platform
google-cloud-vision==3.4.4
google-cloud-translate==3.11.1
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Azure Services
azure-storage-blob==12.19.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-identity==1.15.0

# OpenAI
openai==1.3.7

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching and Session Management
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-google-cloud==1.21.0

# Security
cryptography==41.0.7
python-multipart==0.0.6

# Local fallbacks
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
reportlab==4.0.7
langdetect==1.0.9

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
